+++ lib/rules_autofresh/rules_hotreload.dart
@@ -55,7 +55,7 @@ class _RulesHotReloadState extends ConsumerState<RulesHotReload> {
     fileChanges.close();
     final modVariants = ref.watch(AppState.modVariants).valueOrNull;
     final modRulesDotCsvFiles = modVariants
-            ?.map((variant) => getRulesCsvInModFolder(variant.modsFolder))
+            ?.map((variant) => getRulesCsvInModFolder(variant.modFolder))
             .whereNotNull()
             .toList() ??
         [];