+++ lib/trios/constants.dart
@@ -33,9 +33,9 @@ class Constants {
 
   //     // Backwards compat, first one is the one used for new disable actions.
   static const modInfoFileDisabledNames = [
+    "mod_info.json.disabled",
     "mod_info.json.disabled-by-TriOS",
     "mod_info.json.disabled-by-SMOL",
-    "mod_info.json.disabled"
   ];
 
   static const versionCheckerCsvPath = "data/config/version/version_files.csv";